Asset Entity Completed

+++ src/main/java/com/chahat/trading_platform/service/OrderServiceImpl.java
@@ -2,10 +2,7 @@ package com.chahat.trading_platform.service;
 
 import com.chahat.trading_platform.domain.OrderStatus;
 import com.chahat.trading_platform.domain.OrderType;
-import com.chahat.trading_platform.model.Coin;
-import com.chahat.trading_platform.model.Order;
-import com.chahat.trading_platform.model.OrderItem;
-import com.chahat.trading_platform.model.User;
+import com.chahat.trading_platform.model.*;
 import com.chahat.trading_platform.repository.OrderItemRepository;
 import com.chahat.trading_platform.repository.OrderRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -26,6 +23,9 @@ public class OrderServiceImpl implements OrderService {
 
     @Autowired
     private OrderItemRepository orderItemRepository;
+
+    @Autowired
+    private AssetService assetService;
 
     @Override
     public Order createOrder(User user, OrderItem orderItem, OrderType orderType) {
@@ -73,7 +73,16 @@ public class OrderServiceImpl implements OrderService {
         walletService.payOrderPayment(order, user);
         order.setOrderStatus(OrderStatus.SUCCESS);
         order.setOrderType(OrderType.BUY);
-        return orderRepository.save(order);
+        Order savedOrder = orderRepository.save(order);
+
+        Asset oldAsset = assetService.findAssetByUserIdAndCoinId(order.getUser().getId(), order.getOrderItem().getCoin().getId());
+        if (oldAsset == null){
+            assetService.createAsset(user, orderItem.getCoin(), orderItem.getQuantity());
+        } else {
+            assetService.updateAsset(oldAsset.getId(), quantity);
+        }
+        return savedOrder;
+
     }
 
     @Transactional
@@ -83,27 +92,30 @@ public class OrderServiceImpl implements OrderService {
         }
 
         double sellPrice = coin.getCurrentPrice();
+        Asset assetToSell = assetService.findAssetByUserIdAndCoinId(user.getId(), coin.getId());
 
-        double buyPrice = assetToSell.getPrice();
+        if (assetToSell != null) {
+            double buyPrice = assetToSell.getBuyPrice();
+            OrderItem orderItem = createOrderItem(coin, quantity, buyPrice, sellPrice);
+            Order order = createOrder(user, orderItem, OrderType.SELL);
+            orderItem.setOrder(order);
 
-        OrderItem orderItem = createOrderItem(coin,quantity, buyPrice, sellPrice);
-        Order order = createOrder(user, orderItem, OrderType.SELL );
-        orderItem.setOrder(order);
+            if (assetToSell.getQuantity() >= quantity) {
 
-        if (assetToSell.getQuantity() >= quantity){
+                order.setOrderStatus(OrderStatus.SUCCESS);
+                order.setOrderType(OrderType.SELL);
+                Order savedOrder = orderRepository.save(order);
+                walletService.payOrderPayment(order, user);
 
-            order.setOrderStatus(OrderStatus.SUCCESS);
-            order.setOrderType(OrderType.SELL);
-            Order savedOrder = orderRepository.save(order);
-            walletService.payOrderPayment(order, user);
-
-            Asset updatedAsset = assetService.updateAsset(assetToSell.getId(), quantity);
-            if (updatedAsset.getQuantity()*coin.getCurrentPrice() <= 1){
-                assetService.deleteAsset(updatedAsset.getId());
+                Asset updatedAsset = assetService.updateAsset(assetToSell.getId(), -quantity);
+                if (updatedAsset.getQuantity() * coin.getCurrentPrice() <= 1) {
+                    assetService.deleteAsset(updatedAsset.getId());
+                }
+                return savedOrder;
             }
-            return savedOrder;
+            throw new Exception("Insufficient Quantity To Sell");
         }
-        throw new Exception("Insufficient Quantity To Sell");
+        throw new Exception("Asset Not Found");
     }
 
     @Override
